chore: Move components around for better reusage
